requirements.txt
aiohttp==3.7.3
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.3
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.7.0rc1
Django==3.1.4
django-constance==2.8.0
django-globals==0.3.2
django-timezone-field==4.1.1
django-uuidfield==0.5.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
fcm-django==0.3.7
geoip2==4.1.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.1.1
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
maxminddb==2.0.3
msgpack==1.0.0
multidict==5.1.0
oauthlib==3.1.0
packaging==20.7
prompt-toolkit==3.0.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zipp==3.4.0
zope.interface==5.2.0
